+++ testcontainers-kafka/src/test/java/com/github/ydespreaux/testcontainers/kafka/rule/ConfluentKafkaContainerTest.java
@@ -25,41 +25,42 @@ import com.github.ydespreaux.testcontainers.kafka.config.TopicConfiguration;
 import kafka.admin.AdminUtils;
 import kafka.utils.ZkUtils;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
 @Slf4j
-@RunWith(SpringRunner.class)
-public class ITConfluentKafkaContainerTest {
+@Tag("integration")
+@Testcontainers
+public class ConfluentKafkaContainerTest {
 
-    @ClassRule
+    @Container
     public static final ConfluentKafkaContainer container = new ConfluentKafkaContainer("5.1.0")
             .withSchemaRegistry(true)
             .withTopic("topic1", 3, false)
             .withTopic("topic2-compact", 3, true);
 
     @Test
-    public void containerEnvironment() {
+    void containerEnvironment() {
         assertThat(container.getBootstrapServers(), is(notNullValue()));
         assertThat(container.getZookeeperServer(), is(notNullValue()));
         assertThat(container.getSchemaRegistryServers(), is(notNullValue()));
     }
 
     @Test
-    public void checkTopics() {
+    void checkTopics() {
         ZkUtils zkUtils = ZkUtils.apply(container.getZookeeperServer(), 6000, 6000, false);
         assertThat(AdminUtils.topicExists(zkUtils, "topic1"), is(true));
         assertThat(AdminUtils.topicExists(zkUtils, "topic2-compact"), is(true));
     }
 
     @Test
-    public void createTopic() {
+    void createTopic() {
         container.withTopic(new TopicConfiguration("TOPIC1", 1, false));
 
         ZkUtils zkUtils = ZkUtils.apply(container.getZookeeperServer(), 6000, 6000, false);
@@ -68,7 +69,7 @@ public class ITConfluentKafkaContainerTest {
     }
 
     @Test
-    public void createCompactTopic() {
+    void createCompactTopic() {
         container.withTopic(new TopicConfiguration("TOPIC_COMPACT_1", 1, true));
 
         ZkUtils zkUtils = ZkUtils.apply(container.getZookeeperServer(), 6000, 6000, false);