+++ testcontainers-kafka/src/test/java/com/github/ydespreaux/testcontainers/kafka/rule/ConfluentSecureKafkaConnectContainerTest.java
@@ -21,39 +21,46 @@
 package com.github.ydespreaux.testcontainers.kafka.rule;
 
 
+import com.github.ydespreaux.testcontainers.kafka.CertsDefinition;
 import com.github.ydespreaux.testcontainers.kafka.domain.WorkerInfo;
 import lombok.extern.slf4j.Slf4j;
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.web.client.RestTemplateBuilder;
 import org.springframework.web.client.RestTemplate;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.Matchers.notNullValue;
 
 @Slf4j
-@RunWith(SpringRunner.class)
-public class ITConfluentKafkaConnectContainerTest {
+@Tag("integration")
+@Testcontainers
+public class ConfluentSecureKafkaConnectContainerTest {
 
-    @ClassRule
-    public static final ConfluentKafkaConnectContainer container = new ConfluentKafkaConnectContainer("5.1.0")
-            .withSchemaRegistry(true)
-            .withKeyConverter("org.apache.kafka.connect.storage.StringConverter")
-            .withValueConverter("io.confluent.connect.avro.AvroConverter");
+    @Container
+    public static final ConfluentKafkaConnectContainer container = new ConfluentKafkaConnectContainer()
+            .withKafkaServerCertificates(CertsDefinition.kafkaServerCertificates)
+        .withSchemaRegistry(true)
+        .withKeyConverter("org.apache.kafka.connect.storage.StringConverter")
+        .withValueConverter("io.confluent.connect.avro.AvroConverter");
 
     @Test
-    public void containerEnvironment() {
+    void containerEnvironment() {
         assertThat(container.getBootstrapServers(), is(notNullValue()));
         assertThat(container.getZookeeperServer(), is(notNullValue()));
         assertThat(container.getRestAppServers(), is(notNullValue()));
     }
 
     @Test
-    public void checkAppRest() {
-        RestTemplate template = new RestTemplate();
-        WorkerInfo info = template.getForObject(container.getRestAppServers(), WorkerInfo.class);
+    void checkAppRest() {
+        RestTemplateBuilder builder = new RestTemplateBuilder();
+        builder = builder.rootUri(container.getRestAppServers());
+        RestTemplate template = builder.build();
+
+        WorkerInfo info = template.getForObject("/", WorkerInfo.class);
         assertThat(info, is(notNullValue()));
         assertThat(info.getVersion(), is(notNullValue()));
         assertThat(info.getCommit(), is(notNullValue()));