+++ testcontainers-mysql/src/test/java/com/github/ydespreaux/testcontainers/mysql/MySQLContainerTest.java
@@ -20,20 +20,21 @@
 
 package com.github.ydespreaux.testcontainers.mysql;
 
-import org.junit.ClassRule;
-import org.junit.Test;
-import org.junit.runner.RunWith;
-import org.springframework.test.context.junit4.SpringRunner;
+import org.junit.jupiter.api.Tag;
+import org.junit.jupiter.api.Test;
+import org.testcontainers.junit.jupiter.Container;
+import org.testcontainers.junit.jupiter.Testcontainers;
 
 import static java.lang.String.format;
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalTo;
 import static org.hamcrest.Matchers.is;
 
-@RunWith(SpringRunner.class)
-public class ITMySQLContainerTest {
+@Tag("integration")
+@Testcontainers
+public class MySQLContainerTest {
 
-    @ClassRule
+    @Container
     public static MySQLContainer mySqlContainer = new MySQLContainer()
             .withDatabaseName("my_database")
             .withUsername("db_user")
@@ -48,7 +49,7 @@ public class ITMySQLContainerTest {
             .withStartupTimeoutSeconds(180);
 
     @Test
-    public void environmentSystemProperty() {
+    void environmentSystemProperty() {
         assertThat(System.getProperty(mySqlContainer.getDriverClassSystemProperty()), is(equalTo(mySqlContainer.getDriverClassName())));
         assertThat(System.getProperty(mySqlContainer.getUsernameSystemProperty()), is(equalTo("db_user")));
         assertThat(System.getProperty(mySqlContainer.getPasswordSystemProperty()), is(equalTo("changeme")));
@@ -57,13 +58,13 @@ public class ITMySQLContainerTest {
     }
 
     @Test
-    public void getURL() {
+    void getURL() {
         String url = format("jdbc:mysql://%s:%d/%s", mySqlContainer.getContainerIpAddress(), mySqlContainer.getPort(), mySqlContainer.getDatabaseName());
         assertThat(mySqlContainer.getURL(), is(equalTo(url)));
     }
 
     @Test
-    public void getInternalURL() {
+    void getInternalURL() {
         String url = format("jdbc:mysql://%s:%d/%s", mySqlContainer.getNetworkAliases().get(0), 3306, mySqlContainer.getDatabaseName());
         assertThat(mySqlContainer.getInternalURL(), is(equalTo(url)));
     }